+++ test/shared_memory_check/shmem_segment.hpp
@@ -11,7 +11,7 @@ typedef uint32_t accessCounter_t;
 class shmem_segment_t {
 public:
    shmem_segment_t(int size);
-   shmem_segment_t(const char* _shmemKeyFilePath , int size);
+   shmem_segment_t(const char* _shmemKeyFilePath , size_t size);
    shmem_segment_t(const shmem_segment_t& rhs);
    ~shmem_segment_t();
